move exec to terminal module
