--- src/lib/terminal/index.ts
@@ -1,3 +1,0 @@
-export * from "./filesystem";
-export * from "./types";
-export * from "./utils";
